Add visual error handling to dashboard and detail
Summary: Adds an overlay component that can be embedded into any component in Wikistats. Currently only being used with connectivity problems, once Uniques is merged it should also handle incompatible metrics/families

Test Plan:
Open the dashboard or the detail page with a throttled connection (in developer tools, you can change that under the Network menu).
To test errors, change the url of the metric in the configuration to an invalid value. The error overlay should appear over the metric card or detail page.

Reviewers: mforns

Subscribers: fdans

Differential Revision: https://phabricator.wikimedia.org/D755