Handle long project names in wiki selector
Summary: For long project names and specially for the default 'All project families - All languages', the wiki selector cuts out a part of the text. This patch adjusts the size and font-size of the input component and also adds an animation that scrolls the text when hovering over the input.

Test Plan: The default 'All project families - All languages' should appear uncut in the dashboard view, and trigger a scroll in the detail page view (on hover). If you choose 'Wikipedia - Belarusian (Taraskievica ortography)', it should trigger the animation in both the dashboard view and the detail page.

Reviewers: fdans, Milimetric

Reviewed By: fdans

Maniphest Tasks: T173373

Differential Revision: https://phabricator.wikimedia.org/D785