Fix table view not working with line graphs
Summary:
Ref T170461

Line and bar graphs were handled differently in the table view's prototype. Since I gave a lot of love to the bar chart at the beginning but not so much to the line chart, it was failing to render on the table.

Unifying the logic of the table chart solves this problem.

Reviewers: Milimetric

Reviewed By: Milimetric

Maniphest Tasks: T170461

Differential Revision: https://phabricator.wikimedia.org/D786