Fix x axis in line graph
Summary:
Ref T175265

Until now the x axis was broken for daily granularities and padding and label rotation were outdated, since the changes pushed recently only fixed the bar graph. We should really unify as much logic as possible between the bar chart and the line chart to avoid having to repeat work.

Test Plan: In the detail page, with a line graph, switch to 1-month or 3-month time ranges. The space between the labels in the x axis should be uniform and there should be on overlap between them.

Reviewers: mforns

Reviewed By: mforns

Subscribers: fdans

Maniphest Tasks: T175265

Differential Revision: https://phabricator.wikimedia.org/D788