Add top articles by pageviews metric
This patch changes the list view both in dashboard and detail pages in order
to make tops metrics work. It also adds the configuration for the top articles
by pageviews metric.

In order for the aqs client to properly identify tops metrics, it now has access
to the metrics' configuration.

Bug: T175266
Change-Id: I33e30293d6aa46679bf268b8bfe6db2b9181676b