Fix bug when some of aqs requests return 202 and others 404
The AQS API makes several requests for each metric. For instance,
in the case of pageviews, it makes 3 requests: desktop, mobile
and apps. Normally those 3 will either return 200 all, or return
404 all. However, for smaller wikis, it can happen that some of
them return 200, and some 404, because the site does not have any
app pageviews. In this cases Wikistats, was erroring and not
showing the data that was correct (200). This patch fixes this by
allowing the requests to be completed, even if 404, and then
filtering them out. In the detail view, the breakdowns are
modified to only show the dimension values that have data.

Bug: T176261
Change-Id: I300da49ba869011995e83587db92bf0a7a1ef791