[WIP] Clean up UI and aggregations
This cleans up a few things noticed by Erik in the linked task
It also refactors the aggregation a bit and corrects some calculation
mistakes.  The main thing remaining is to decide what to show on the
Dashboard because YoY numbers are not always going to match the widget
graph visually.

Bug: T178084
Change-Id: Id2bd700f9f72aea6ee56d68234426979d355e04d