Query breakdowns only when they are visualised
Until now we were requesting breakdown data regardless of whether we were
using it or not, under the assumption that each metric only had one possible
breakdown option. This change makes the dashboard and detail pages request
data without breakdown filters unless a breakdown value is selected in the
sidebar.

Bug: T178461
Change-Id: I319a3e7e41d61779d7b1158052f32801e7b5d4a1