Remove download of font from google endpoint
This change splits off the css outside of the bundle for development. Spliting
CSS and JS makes a lot easier the configuration of the fonts. There is a new file (lato.css) that
stores the references to the fonts we are using so they are managed by our code
rather than referencing google.

Font files are embeded or downloaded depending on size.

Bug: T178317
Change-Id: Ifdec8c07f3eb578112b779105f51d0e2781d5965