Simplify and fix breakdowns and other data
Three substantial refactorings of the data structures and related code:

First, there is always a breakdown, so we don't have to have duplicated
code for the breakdown vs. no-breakdown case.  The no-breakdown case is
now just a breakdown with a single value of "total" that's inserted when
the GraphModel is created.

Second, the "activeBreakdown" is now explicitly set instead of being
inferred from the breakdown that has "on" set to true.  This makes it
more natural to sync with the UI and actually inspires the UI to be a
little more intuitive in my opinion (this should be checked by others).

Third, the GraphModel constructor no longer takes data.  The data is set
after the fact when the promise returns.  This allows the Detail to
coordinate changes in breakdowns without re-rendering and deep-watching
the graphModel property.

I don't think this change cleans up the code completely, I'm still being
clumsy about what data's being passed around.  But it does work in
Firefox and I think it simplifies the code.

Bug: T180556
Change-Id: I034b70b3d5a0034aa87013650884691c1c94e9f9