Fix loading sparse data into widgets
When data isn't continuous, widgets behave in hard-to-predict ways like
showing NaNs and incorrect YoY or MoM totals.  This patch addresses most
of that but I'm not really sure what to do when there's data *more* than
a year ago.  The change from the last available datapoint beyond a year
seems relevant, but it's not YoY change.

Bug: T182224
Change-Id: If911e6dceeb9281887e19e705ae716c1ea1b6772