Fix infinite loop
For the future, the flow of events and data in WikiSelector is too
circular.  The init calls get triggered when they're not necessary,
after state changes.

Also, the way the hidden div is added with an id is not great, that
should at least have a class and in general should be added relative to
the component in case multiple versions of the same component are used
on the page.

Bug: T182700
Change-Id: I1dd5f4bce62caaf14215b4fa158d030f53e8f084