Replacing JSON download with CSV download
While ticket asks to add a csv download option I think it makes sense
to have just one download option and have that be in csv format.

In order to convert seamlessly from json to csv the json data is flattened
before being passed to d3 to do the conversion.

Bug: T183192
Change-Id: I879356fd3d3f197efa51d4c16b5e27a1cccc153e