Optimize WikiSelector for slow browsers
WikiSelector can have more than 1000 suggested options,
and each one of them contains several html elements.
In some browsers, the rendering of such many DOM nodes
is very slow and takes seconds before options show up.

This fix loads the search results in chunks of 50.
It loads the first 50, and then when the rendering is done,
proceeds to load the subsequent 50, and so on.
The user can already start interacting with the control
after the first (quick) load, and does not have to wait
for all the hidden options to render.

Also, small caching is added so clearing of search text in selector
doesn't repeat computations.

Bug: T185334
Change-Id: Ia81a7445c3715064518026ccc1f149ff230770f0