Annotate wikistats
A new annotation api fetches annotations from mediawiki and transforms
them into an array.  The Detail component sets a promise on the graph
model with the result of an annotation fetch.  The graphs themselves
then add a callback to the promise to render when annotations are
available.  This last part must be done in graphs to give them a chance
to group overlapping annotations, and adjust rendering so annotations
don't overflow outside the paint area.

Features for the Future
* implement annotations on table charts
* implement date range annotations, maybe using a different visual style
* show annotations for a single point in a scrollable popup if the text is too big

Bug: T194705
Change-Id: I56796a45e2f7200dabea218b3d9a69a0e10c26b4