Add pages to date metric
This change implements the capability to transform datasets by
applying a function to them before they are passed to the graphs.

We're adding the total article count metric by taking the new pages
endpoint and applying a cumulative function to the data.

This also fixes a rendering problem with constant horizontal lines
and the top of the line charts in the dashboard, and renames the
"legacy" configuration property to "frozen".

Bug: T203180
Change-Id: Iee20c7f380c12e0e8085888d06994fc45da7cb2a