Fix last 12 months aggregate
Until now we were taking the first 12 months of the coming dataset,
which was pretty difficult to spot because with non-frozen metrics
we would get data from the last year.

With small wikis on frozen datasets though, this bug is made obvious
by the fact that the 12-month variation shows zero (see Bambara Wiki).

Bug: T205565
Change-Id: Ica7758cedc9c63a88b21d30ee16677f8f8e4af1b