Allow several attempts to get latest data in top metrics
Since in top metrics we need a concrete date to fetch data and we
don't have a way to fetch our data's ranges, this change adds the
ability of the AQS client to re-form the request and try again, up
to 3 times.

This change also removes the short-lived but very honorable
availabilityBuffer property.

Goodnight, sweet prince, and flights of angels sing thee to thy rest.

Bug: T205915
Change-Id: I01ddf77d14541e838011908a6ffa1e815e952333