Add tooltip Vue directive and use it
This change cleans up a little clutter and organizes directives and
filters in separate folders.  It then adds a hint directive to be used
for accessible tooltips.  All tooltip text is factored out into a config
file and imported async via webpack's chunking system so it doesn't
increase the initial page load time.

Bug: T177950
Change-Id: Idfcfe05598c266431129b480395e3658ad8e4741