Create metrics matrix component
This is a resubmission of the accidentally merged patch, for further code review.

This component shows all Wikistats metrics with tooltips explaining each,
restricts navigation to them when uncompatible project is selected and signals
the way to consult Wikistats 1 metrics.

The routing part of wikistats has been changed to accommodate further sections
other than just the detail page and the dashboard.

I've moved a bunch of css between components to enforce scoped styles.

Bug: T187806
Change-Id: I58e8e7daa094a02e235e4ce6087416a9e22840b4