Filter out unavailable metrics on desktop
Before, the availableMetrics computed was ignoring unavailable metrics
only on mobile.  Now, it makes two lists and uses as many available
metrics as possible until it runs out, then fills in the
availableMetricSlots with unavailable metrics.  This makes the order of
metrics more predictable and allows the 2-slot layout a chance to show
available metrics only.  I made the availableMetricSlots a computed,
which as a side-effect seems to have stabilized some of the
resize/reflow problems when toggling between 1, 2, 3-slot layouts and
mobile layout.  I think it just lets Vue have more control over when to
do that stuff more optimally but I haven't looked into it.

Bug: T220083
Change-Id: I2d1ba13ccf98f2f3e003a03b1e7286f75d5569cb