Use current date as end date in relative time ranges
With the new time selector we've been using the last day of the
previous month as end date, which makes druid metrics return one
month less than expected. While we work on unifying the time range
formats, we're setting the end date as the current date to make sure
we always get the most recent available data.

Bug: T226421
Change-Id: Ie0cbe9a5023b9d9c50922d87e3edb5f77ff11bbe