Add zeroes to truncated values and UI about truncation
Right now the graphs are skipping truncated values in uniques.
This looks particularly weird in many small wikis in daily granularity.
This change fills those values with zeroes so that they can be
rendered as truncated. It also adds a "truncated" property so that
we can handle them in the charts.

Graphs now have a truncated area that explains why values under 1000 are
absent.

Bug: T220098
Change-Id: I04bd08ade41eb20acbc61ed149e74d76b00d2868