Tell map to render on chart type changes
The map chart is a little different than the others, because it loads
some of its dependencies async via a promise.  Because of this, it has
to make a computed on multiple promises.  If the graphData is already
loaded when switching charts, the computed won't fire for it.  So it
won't compute dataByCountry.  I added a note to this effect, please -1
if it's not clear because I was pretty lost there for a bit trying to
figure it out.

Bug: T226514
Change-Id: Ia1638da1ce70d8e8f18f280e9a2864c180b6d26a