Remove unnecessary child reference
One of the things that made this a bit harder to debug is this
unnecessary parent hook from GraphPanel into Bar and Line.  This is not
needed and I regret adding it, especially since I added it at the same
time as the 'detail/fullscreen' store property which renders it useless.

Bug: T226514
Change-Id: I01adc68fc3c0f6033f223715a9fde0c35383977b