Fix end date decay bug
There were a lot of unrelated bugs on the path to this one.  Basically,
when data came back from AQS, the time range end would be set to the
last month of data, which we treat as exclusive in AQS, so the next
refresh would get one less month.  This is cyclical and so every time
you refresh the data, whether because you select a different project or
resize the screen, the end date would decay by one month.  The "feature"
was there I believe to correctly show graphs without data in the latest
months.  I'm not sure what to do about that corner case right now, but I
am submitting this fix to the more serious bug.

Bug: T230514
Change-Id: If82e55f27ac3a0ce8ef7d43c994e708e36f63cfb