Transition data rows to using time ranges instead of timestamps
The bit of code that was adjusting the time range of the metric
according to the data received was using the last row's timestamp
as a comparison, which is wrong because until now the timestamp has
always been the beginning of the row's period. This change adds a
timerange field to the rows and factors out some of the uses of
the timestamp.

Eventually it should be easy to get rid of timestamp altogether, but
for now I didn't feel it was necessary to do the whole transition.

Bug: T230514
Change-Id: I79b7544b6111b1d80d99529a66ea67e1924f5a82