Use different logic for chart axis number formatting from dashboard
The priority when showing numbers in the dashboard is giving a clear
picture of the metric without overwhelming the user with unnecessary
digits. This logic, however, when applied to axes in the charts, makes
the numbers aggressively round like in the example of this change's
task.

This change makes the line and bar charts use a slightly different
numeral filter than the dashboard to address this problem.

Bug: T242790
Change-Id: I7dc5b3b9eeaf51dd02dbc56ee37c3a3aed43e04a