Don't enable locales if they haven't been built
This change corrects a bug that made the website load a locales that
had been pushed by TranslateWiki but hadn't been built yet (generally
because the locale was incomplete so we chose not to build it).

To reproduce, enter the website and change manually the localStorage
property "language" to nl (one of the languages that didn't make the
cut in the last deployment), and the site should give a 404.

Bug: T246976

Change-Id: Ie7bb58a733762cedc3369112553d5b760e96e4d9