Replace numeral with numbro and fix bytes formatting
Numeral has a bug in the way it was formatting bytes, and since
the library has stopped being maintained, we're replacing it
with numbro. This change fixes the bug in bytes metrics.

Since Numbro has yet another format for number locales, we're
moving to a more explicit list of available languages, removing the
`languages` property in package.json and creating a languages.json
file in src. This also has the nice benefit of solving the sub-languages
problems we had with using the sitematrix as our language base.

Bug: T199386
Change-Id: Ie5a98e72f520c13f8a75061d9d0b6a104810555a