Fix time range decay problem
This fix was reverted in the past and we can't remember why, so we're
deploying it again.  It's better than the current situation, where the
time selector just moves back by a month anytime graph data is changed.
The better fix is to clean up the data flow so nobody ever updates the
store state outside of a commit call.  This refactor will take longer,
so we are just putting this fix out for now.

Bug: T253861
Change-Id: Id23c76e20453d2f9a889082eab058189ce5ee02f