Allow more than one dimension to be filtered in Wikistats
This change phases out the long-standing assumption that there
is only one active "breakdown" in a given graph. The new behavior
will only accept one splitting dimension (what we previously
referred to as "breakdown"), but allow more than one dimension to
be filtered by.

Additionally, this change also allows metrics to be filtered with
and without splitting, meaning more complex questions that involve
aggregation can now be answered with Wikistats (e.g. what's the
evolution of the total amount of user editors, by number of
content-only edits, with a higher edit count than 5 per month?).

Since it's a pretty big change in the inner workings of the app I've
tried to add tests as much as possible and rewrite a couple of functions
with proper testing included.

Bug: T255757
Change-Id: I40d5a676422b167bcfddc7d69b5f431bbac3fbe5