Consider locked dimensions when exploding keys for AQS url
This patch corrects a bug that was causing the based-on metrics
(metrics that inherit other metrics, with different dimension values)
to ignore the "locked" setting on the dimensions settings, therefore
using the same data as the metrics they are based on. This change fixes
the user edits and active editors metrics.

Bug: T265322
Change-Id: If3697c553f52d3ce630a6f08198ebc7b5ca88b3d