Add Active Editors per Country metric to Wikistats
Apart from adding the metric configuration, this change adapts the UI to
the caveats in the metric, most notably the absence of an "all" value in
the activity level dimension, which means radio buttons must spawn instead
of checkboxes when filtering the metric. This change also includes a small
UX improvement that will hide the dimension selector card when there's only
one selectable dimension, opening it in the first place instead.

Bug: T188859
Change-Id: Ica75ebe8b0d2690a66a0b98bb27934d90f211780