Remove metric groups
This feature was added to help with metric discoverability.  But it
caused a bug and was hard to fix without either adding tech debt or
paying down a lot of older tech debt.  So we're disabling it for now,
and keeping metric discoverability in mind for our next big refresh.

Bug: T262725
Change-Id: I4431b855bbc03ee613b83887c64bca766949198c