Alter routing logic to allow value lists
This change only adds the ability for the router to accept more
than one value for wiki projects and metrics. The metric and project
values are parsed as arrays but for now they are not propagated into
the state, instead being reduced back into the first element of the array.

Bug: T283596
Change-Id: Iae6b3a0666b50f242d6635c313ab9d93ab11cdfc