Correct pageview data description
The description was incorrectly implying that the metric is only showing
human user pageviews.  By default, this is not true, we include
automated pageviews.  Hopefully this change clarifies things, apologies
for the translation confusion that this causes (I take responsibility if
this creates a mess, any advice on how to help correct it is
appreciated).

Bug: T295298
Change-Id: I73f89427ab7ff08c59ef56d04e078973ca6c62cc