Link to AQS documentation instead of Research page
The "Reading" metrics on wikistats were linking to the generic
Research:Page_view page instead of the more specific wikitech
documentation on AQS.  This updates the AQS documentation with some
details on agent types (namely, only the "user" agent types are
considered in the pageviews by country and most read articles metrics).
It then links to these updated definitions from wikistats in a way that
doesn't require translation.  This change could have added some
clarifying text to the metric descriptions, but with those likely to
continue changing it feels safer to link out and include any detail on
wikitech.  This also, of course, keeps those details from being
translated, so we may have to revisit.

Bug: T295298
Change-Id: I2ba1adcc96f4fe356edac7f729620b7c852eda01