Add Dockerfile to build production dist folder
This repository checks in the `dist` directory for deployment, but we want
to create the dist directory in docker for reproducible builds.

Thus to use this `Dockerfile`, attach a volume to the `dist` directory.
This mirrors the changes in docker so the resulting `dist` can be committed
once the build finishes:

    export WIKISTATS_VERSION=$(jq -r .version package.json)
    docker build . -t wikistats2:$WIKISTATS_VERSION
    docker run --volume $(pwd)/dist:/app/dist wikistats2:$WIKISTATS_VERSION

At this point the docker image wikistats2-2.9.3 or whatever could be
pushed to a container registry. Hopefully someday we'll deploy wikistats
via container that way or some other way that does not involve checking
`node_modules` in to git.

If you'd like to iterate on a single app version, I recommend manually
adding a build version suffix as follows:

    docker build . -t wikistats2:$WIKISTATS_VERSION_0

Or simply build an anonymous image and run the last one:

    docker build .
    docker run --volume (pwd)/dist:/app/dist $(docker images -q | head -1)

Change-Id: Ibd0cfa8f371b404bff9a833c93ed7dc07ea5f2c0