Avoid infinite loop when loading pages-to-date
When loading this metric, various points in the data flow were looking
at DimensionalData::getAllItems.  This was happening just to check the
size of the returned dataset.  And that's ok, it's useful for wikis that
have sparse data so that wikistats can step backwards in time and find
the first month that returns something.  But it's better to check the
size of the result and not get all the items themselves.  Because
crossfilter was sorting the results when queried this way.  It uses
recursive quicksort and with a large dataset that's more or less in
order, I think it was recursing more than the allowed stack depth.

Simple mitigation:

1. implement isEmpty in DimensionalData and use it instead of
   getAllItems().length === 0
2. replace the usage of getAllItems for size checking

Additional fix: add some missing tooltips
Additional fix: add some clarifying comments

Bug: T312717
Change-Id: I2f749bbcab5c2ba04476a7a6c599950bb9f67dfa