Increase world map resolution
This hurts performance by about 600Kb, because the new topojson is much
bigger.  However, even on 3G this is only a few seconds worth of penalty
and we get 70 additional countries to report on.

NOTE: the countries-50m topojson does not need to be loaded async
because the whole MapChart component is already loaded async from
Detail.vue

Bug: T338033
Change-Id: Ib15432501accde86c19d54c9e4c9f892a796015e